chore: extract functions related to static analysis from index.ts
